+++ P2/src/util.cpp
@@ -13,36 +13,38 @@
 
 using namespace std;
 
-Result Free2DArray(unsigned int** array, size_t n) {
-    for (size_t i = 0; i < n; i++)
+Result Free2DArray(unsigned** array, unsigned n) {
+    for (unsigned i = 0; i < n; i++)
             free(array[i]);
     free(array);
     return SUCCESS;
 }
 
-size_t Fact(size_t n) {
-    size_t ret = 1;
-    for(size_t i = 1; i <= n; i++)
+/* TODO: add exception. */
+unsigned Fact(unsigned n) {
+    unsigned ret = 1;
+    for(unsigned i = 1; i <= n; i++)
         ret *= i;
     return ret;
 }
 
-void print_vector(vector<size_t> v) {
+void print_vector(vector<unsigned> v) {
     for (auto const& c : v)
         cout << c << ' ';
     cout << endl;
 }
 
-size_t subtract(size_t a, size_t b) {
+/* TODO: add exception. */
+unsigned subtract(unsigned a, unsigned b) {
     if (a < b)
         cout << "Util: Warning subtract lead to underflow" << endl;
     return a - b;
 }
 
-bool between_right(size_t a, size_t b, size_t c) {
+bool between_right(unsigned a, unsigned b, unsigned c) {
     return ((a <= b) && (b < c));
 }
 
-bool between_left(size_t a, size_t b, size_t c) {
+bool between_left(unsigned a, unsigned b, unsigned c) {
     return ((a < b) && (b <= c));
 }
